V2.63.6
Muchos cambios

+++ modelo/wada.php
@@ -11,8 +11,15 @@ class WADA extends datos
 
     public function incluir_wada($id_atleta, $estado, $inscrito, $ultima_actualizacion, $vencimiento)
     {
-        if (!Validar::validar("cedula", $id_atleta)) {
+        if (!Validar::validar("cedula", $id_atleta)["ok"]) {
             return ["ok" => false, "mensaje" => "La cedula del atleta no es valida"];
+        }
+        if (!Validar::validar("bool", $estado)["ok"]) {
+            return ["ok" => false, "mensaje" => "El estado de la WADA no es valido"];
+        }
+        $validar_fechas = Validar::validar_fechas_wada($this->conexion, $id_atleta, $inscrito, $ultima_actualizacion, $vencimiento);
+        if (!$validar_fechas["ok"]) {
+            return ["ok" => false, "mensaje" => $validar_fechas["mensaje"]];
         }
         $this->id_atleta = $id_atleta;
         $this->estado = $estado;
@@ -24,6 +31,16 @@ class WADA extends datos
 
     public function modificar_wada($id_atleta, $estado, $inscrito, $ultima_actualizacion, $vencimiento)
     {
+        if (!Validar::validar("cedula", $id_atleta)["ok"]) {
+            return ["ok" => false, "mensaje" => "La cedula del atleta no es valida"];
+        }
+        if (!Validar::validar("bool", $estado)["ok"]) {
+            return ["ok" => false, "mensaje" => "El estado de la WADA no es valido"];
+        }
+        $validar_fechas = Validar::validar_fechas_wada($this->conexion, $id_atleta, $inscrito, $ultima_actualizacion, $vencimiento);
+        if (!$validar_fechas["ok"]) {
+            return ["ok" => false, "mensaje" => $validar_fechas["mensaje"]];
+        }
         $this->id_atleta = $id_atleta;
         $this->estado = $estado;
         $this->inscrito = $inscrito;
@@ -34,12 +51,18 @@ class WADA extends datos
 
     public function obtener_wada($id_atleta)
     {
+        if (!Validar::validar("cedula", $id_atleta)) {
+            return ["ok" => false, "mensaje" => "La cedula del atleta no es valida"];
+        }
         $this->id_atleta = $id_atleta;
         return $this->obtener();
     }
 
     public function eliminar_wada($id_atleta)
     {
+        if (!Validar::validar("cedula", $id_atleta)) {
+            return ["ok" => false, "mensaje" => "La cedula del atleta no es valida"];
+        }
         $this->id_atleta = $id_atleta;
         return $this->eliminar();
     }
@@ -52,8 +75,17 @@ class WADA extends datos
     private function incluir()
     {
         try {
+            $consulta = "SELECT cedula FROM atleta WHERE cedula = ?;";
+            $existe = Validar::existe($this->conexion, $this->id_atleta, $consulta);
+            // Se verifica que el atleta exista para asignarle la WADA
+            if (!$existe["ok"]) {
+                $resultado["ok"] = false;
+                $resultado["mensaje"] = "Este atleta no existe";
+                return $resultado;
+            }
             $consulta = "SELECT id_atleta FROM wada WHERE id_atleta = ?;";
             $existe = Validar::existe($this->conexion, $this->id_atleta, $consulta);
+            // Se verifica que no exista la WADA para este atleta
             if ($existe["ok"]) {
                 $resultado["ok"] = false;
                 $resultado["mensaje"] = "Ya existe la WADA de este atleta";
@@ -86,6 +118,14 @@ class WADA extends datos
     private function modificar()
     {
         try {
+            $consulta = "SELECT id_atleta FROM wada WHERE id_atleta = ?;";
+            $existe = Validar::existe($this->conexion, $this->id_atleta, $consulta);
+            if (!$existe["ok"]) {
+                $resultado["ok"] = false;
+                $resultado["mensaje"] = "No existe la WADA de este atleta";
+                return $resultado;
+            }
+            $this->conexion->beginTransaction();
             $consulta = "UPDATE wada SET estado = :estado, inscrito = :inscrito, ultima_actualizacion = :ultima_actualizacion, vencimiento = :vencimiento 
                          WHERE id_atleta = :id_atleta";
             $valores = array(
@@ -97,41 +137,65 @@ class WADA extends datos
             );
             $respuesta = $this->conexion->prepare($consulta);
             $respuesta->execute($valores);
+            $respuesta->closeCursor();
+            $this->conexion->commit();
             $resultado["ok"] = true;
-        } catch (Exception $e) {
+        } catch (PDOException $e) {
+            $this->conexion->rollBack();
             $resultado["ok"] = false;
             $resultado["mensaje"] = $e->getMessage();
         }
+        $this->desconecta();
         return $resultado;
     }
 
     private function obtener()
     {
         try {
+            $consulta = "SELECT id_atleta FROM wada WHERE id_atleta = ?;";
+            $existe = Validar::existe($this->conexion, $this->id_atleta, $consulta);
+            if (!$existe["ok"]) {
+                $resultado["ok"] = false;
+                $resultado["mensaje"] = "No existe la WADA de este atleta";
+                return $resultado;
+            }
             $consulta = "SELECT * FROM wada WHERE id_atleta = :id_atleta";
             $respuesta = $this->conexion->prepare($consulta);
             $respuesta->execute([':id_atleta' => $this->id_atleta]);
             $respuesta = $respuesta->fetch(PDO::FETCH_ASSOC);
             $resultado["ok"] = true;
             $resultado["wada"] = $respuesta;
-        } catch (Exception $e) {
+        } catch (PDOException $e) {
             $resultado["ok"] = false;
             $resultado["mensaje"] = $e->getMessage();
         }
+        $this->desconecta();
         return $resultado;
     }
 
     private function eliminar()
     {
         try {
+            $consulta = "SELECT id_atleta FROM wada WHERE id_atleta = ?;";
+            $existe = Validar::existe($this->conexion, $this->id_atleta, $consulta);
+            if (!$existe["ok"]) {
+                $resultado["ok"] = false;
+                $resultado["mensaje"] = "La WADA del atleta ingresado no existe";
+                return $resultado;
+            }
+            $this->conexion->beginTransaction();
             $consulta = "DELETE FROM wada WHERE id_atleta = :id_atleta";
             $respuesta = $this->conexion->prepare($consulta);
             $respuesta->execute([':id_atleta' => $this->id_atleta]);
+            $respuesta->closeCursor();
+            $this->conexion->commit();
             $resultado["ok"] = true;
-        } catch (Exception $e) {
+        } catch (PDOException $e) {
+            $this->conexion->rollBack();
             $resultado["ok"] = false;
             $resultado["mensaje"] = $e->getMessage();
         }
+        $this->desconecta();
         return $resultado;
     }
 
@@ -142,7 +206,6 @@ class WADA extends datos
                     u.cedula, 
                     u.nombre, 
                     u.apellido, 
-                    a.entrenador,
                     w.estado,
                     w.inscrito,
                     w.vencimiento,
@@ -156,10 +219,11 @@ class WADA extends datos
             $respuesta = $respuesta->fetchAll(PDO::FETCH_ASSOC);
             $resultado["ok"] = true;
             $resultado["respuesta"] = $respuesta;
-        } catch (Exception $e) {
+        } catch (PDOException $e) {
             $resultado["ok"] = false;
             $resultado["mensaje"] = $e->getMessage();
         }
+        $this->desconecta();
         return $resultado;
     }
 
@@ -191,10 +255,11 @@ class WADA extends datos
             $resultado["ok"] = true;
             $resultado["devol"] = 'listado_atletas';
             $resultado["respuesta"] = $respuesta;
-        } catch (Exception $e) {
+        } catch (PDOException $e) {
             $resultado["ok"] = false;
             $resultado["mensaje"] = $e->getMessage();
         }
+        $this->desconecta();
         return $resultado;
     }
 
@@ -202,24 +267,25 @@ class WADA extends datos
     {
         try {
             $consulta = "SELECT u.cedula, 
-                    u.nombre, 
-                    u.apellido, 
-                    w.inscrito,
-                    w.vencimiento,
-                    w.ultima_actualizacion
-                FROM atleta a
-                INNER JOIN usuarios u ON a.cedula = u.cedula
-                INNER JOIN wada w ON w.id_atleta = u.cedula
-                WHERE w.vencimiento <= DATE_ADD(CURDATE(), INTERVAL 30 DAY) ORDER BY vencimiento DESC;";
+                u.nombre, 
+                u.apellido, 
+                w.vencimiento
+            FROM atleta a
+            INNER JOIN usuarios u ON a.cedula = u.cedula
+            INNER JOIN wada w ON w.id_atleta = u.cedula
+            WHERE w.vencimiento > CURDATE() 
+            AND w.vencimiento <= DATE_ADD(CURDATE(), INTERVAL 30 DAY)
+            ORDER BY w.vencimiento DESC;";
             $respuesta = $this->conexion->prepare($consulta);
             $respuesta->execute();
             $registros = $respuesta->fetchAll(PDO::FETCH_ASSOC);
             $resultado["ok"] = true;
             $resultado["respuesta"] = $registros;
-        } catch (Exception $e) {
+        } catch (PDOException $e) {
             $resultado["ok"] = false;
             $resultado["mensaje"] = $e->getMessage();
         }
+        $this->desconecta();
         return $resultado;
     }
